#39 Fix bound function arity calculation based on bound arguments count

+++ src/main/java/com/shade/lang/runtime/objects/function/Function.java
@@ -92,7 +92,7 @@ public abstract class Function extends ScriptObject {
         return name;
     }
 
-    public int getArity() {
+    public byte getArity() {
         return arity;
     }
